refactor: move instance access out of path.py
Path carried two halves that shared nothing but the parsed form: what a path
designates on a model, and how it reads and writes an instance. The second
now lives in its own module, as functions taking the path they act for, and
Path keeps its declarative surface next to the three methods that delegate
to them.
